sepreate book class with person class
